hacc: fix bug where numThreads were different in places
nthreads and numThreads are used in different places in HACC.  However,
it appears that numThreads > nthreads causes segfaults because the iq
vector is accessed out of bounds.  To resolve this, create a new global
variable with the number of threads and use this for both numThreads and
nthreads.  I tested this worked with 4 threads.

Sidenote: although HACC refers to these as threads, they appear to
actually be GPU streams.