resources: Update disk image source to add `hypercall` and `gem5bridge-driver` (#69)
This PR updates the disk image build source for the X86 and Arm base Ubuntu and NPB disk images. The built disk images will now use hypercalls instead of classic exits, and will now also use the gem5 bridge driver, which allows for `gem5-bridge` commands to be run without the use of `sudo`. Both hypercalls and the gem5 bridge driver were introduced in gem5 v25.0.

The build process for the base Ubuntu disk images supports Ubuntu 24.04 and 22.04, and the build process for the NPB disk images supports Ubuntu 24.04 only.